# perf: remove argon bench on boot, use sane defaults across the board