--- a/lerna.json
+++ b/lerna.json
@@ -1,6 +1,6 @@
 {
   "lerna": "5.2.0",
-  "version": "1.52.12",
+  "version": "1.52.13",
   "npmClient": "yarn",
   "useWorkspaces": true,
   "command": {