--- a/packages/hdwallet-keplr/package.json
+++ b/packages/hdwallet-keplr/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shapeshiftoss/hdwallet-keplr",
-  "version": "1.52.12",
+  "version": "1.52.13",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
@@ -15,7 +15,7 @@
   },
   "dependencies": {
     "@shapeshiftoss/caip": "8.15.0",
-    "@shapeshiftoss/hdwallet-core": "1.52.12",
+    "@shapeshiftoss/hdwallet-core": "1.52.13",
     "@shapeshiftoss/proto-tx-builder": "^0.8.0",
     "@shapeshiftoss/types": "3.1.3",
     "base64-js": "^1.5.1",