--- a/packages/hdwallet-ledger-webusb/package.json
+++ b/packages/hdwallet-ledger-webusb/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shapeshiftoss/hdwallet-ledger-webusb",
-  "version": "1.52.12",
+  "version": "1.52.13",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
@@ -20,8 +20,8 @@
     "@ledgerhq/hw-transport-webusb": "^6.7.0",
     "@ledgerhq/live-common": "^21.8.2",
     "@ledgerhq/logs": "^6.10.1",
-    "@shapeshiftoss/hdwallet-core": "1.52.12",
-    "@shapeshiftoss/hdwallet-ledger": "1.52.12",
+    "@shapeshiftoss/hdwallet-core": "1.52.13",
+    "@shapeshiftoss/hdwallet-ledger": "1.52.13",
     "@types/w3c-web-usb": "^1.0.4",
     "p-queue": "^7.4.1"
   },