--- a/packages/hdwallet-trezor-connect/package.json
+++ b/packages/hdwallet-trezor-connect/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shapeshiftoss/hdwallet-trezor-connect",
-  "version": "1.52.12",
+  "version": "1.52.13",
   "license": "MIT",
   "publishConfig": {
     "access": "public"
@@ -14,8 +14,8 @@
     "prepublishOnly": "yarn clean && yarn build"
   },
   "dependencies": {
-    "@shapeshiftoss/hdwallet-core": "1.52.12",
-    "@shapeshiftoss/hdwallet-trezor": "1.52.12",
+    "@shapeshiftoss/hdwallet-core": "1.52.13",
+    "@shapeshiftoss/hdwallet-trezor": "1.52.13",
     "@trezor/rollout": "^1.2.0",
     "trezor-connect": "^8.2.1"
   }