## Overview

Hardcodes sensible defaults for argon encryption parameters, skipping the benchmark to reduce initial app boot time by >4s on both desktop and mobile devices, when connecting to native wallet for the first time.

The default parameters remain the same, with only the `iterations` changing from 23 to 16. Noting that reasonably modern mobile devices achieve ~11 iterations within the 1s target, the default was set to a conservative 16 iterations as a balance between performance and security. 

## Testing

This should be tested locally with shapeshift web via local npm registry (e.g [verdaccio](https://verdaccio.org/))
- **can open native wallet with existing local state** (existing parameters set in IndexedDB with prior benchmark on prior version version). This should prove backwards compatibility so users don't have to re-import their wallets
- **can open native wallet starting from empty local state** (clear IndexedDB)

## Issue

Will close [#5557](https://github.com/shapeshift/web/issues/5557) after bumping version in web.

### Benchmark parameters

**Mobile - Pixel 4a**
Note overall duration of the bench was 4.1s
![image](https://github.com/shapeshift/hdwallet/assets/125113430/521ecd10-bcbf-4600-ad16-65627050a808)

**Desktop - Mac M2**
Note overall duration of the bench was 4.6s
![image](https://github.com/shapeshift/hdwallet/assets/125113430/b792caf4-1751-4d25-a1f1-756c40e32ce8)

### Performance results (on Desktop - Mac M2)

**Before changes**
4s benchmark runs on first open of native wallet:
![image](https://github.com/shapeshift/hdwallet/assets/125113430/1b6ed582-91d9-4505-8c1e-6e325c3bc4bb)

**After changes**
4s benchmark no longer run:
![image](https://github.com/shapeshift/hdwallet/assets/125113430/20406ab8-8209-421b-ab85-9492939fd65e)

Note the defaults get stored as intended:
![image](https://github.com/shapeshift/hdwallet/assets/125113430/0a4a99ca-7783-4d4e-b382-5ce8aa2b83e5)