@gomesalexandre 2023-11-28
Current dependencies on/for this PR:
* master
  * **PR #654** <a href="https://app.graphite.dev/github/pr/shapeshift/hdwallet/654?utm_source=stack-comment-icon" target="_blank"><img src="https://static.graphite.dev/graphite-32x32-black.png" alt="Graphite" width="10px" height="10px"/></a>  👈

This [stack of pull requests](https://stacking.dev/?utm_source=stack-comment) is managed by [Graphite](https://app.graphite.dev/github/pr/shapeshift/hdwallet/654?utm_source=stack-comment).